Fixed a minor bug in leaderboards screen
